info

+++ python/sglang/srt/speculative/build_eagle_tree.py
@@ -46,7 +46,7 @@ __global__ void build_tree(Tensor<long, 2> parent_list, Tensor<long, 2> selected
                 break;
             }
 
-            int token_idx = parent_list[bid][parent_tb_idx];
+            long token_idx = parent_list[bid][parent_tb_idx];
             for(cur_position=0; cur_position<draft_token_num;cur_position++){
                 printf("check equal: left=%d, right=%ld, equal=%d\\n", selected_index[bid][cur_position], token_idx,selected_index[bid][cur_position]==token_idx?1:0);
                 if(selected_index[bid][cur_position]==token_idx){
